Move ObservationChecker to gammapy/data/observations.py
