+++ gammapy/data/data_store.py
@@ -9,7 +9,7 @@ from ..utils.scripts import make_path
 from ..utils.testing import Checker
 from .obs_table import ObservationTable
 from .hdu_index_table import HDUIndexTable
-from .observations import DataStoreObservation, ObservationList
+from .observations import DataStoreObservation, ObservationList, ObservationChecker
 
 __all__ = [
     'DataStore',
@@ -489,154 +489,3 @@ class DataStoreChecker(Checker):
             for records in ObservationChecker(obs).run():
                 yield records
 
-
-class ObservationChecker(Checker):
-    """Check an observation.
-
-    Checks data format and a bit about the content.
-    """
-
-    CHECKS = OrderedDict([
-        ('events', 'check_events'),
-        ('gti', 'check_gti'),
-        ('aeff', 'check_aeff'),
-        ('edisp', 'check_edisp'),
-        ('psf', 'check_psf'),
-    ])
-
-    def __init__(self, obs):
-        self.obs = obs
-
-    def _record(self, level='info', msg=None):
-        return {
-            'level': level,
-            'obs_id': self.obs.obs_id,
-            'msg': msg,
-        }
-
-    def check_events(self):
-        yield self._record(level='debug', msg='Starting events check')
-
-        try:
-            events = self.obs.load('events')
-        except:
-            yield self._record(level='warning', msg='Loading events failed')
-            return
-
-        if len(events.table) == 0:
-            yield self._record(level='error', msg='Events table has zero rows')
-
-    def check_gti(self):
-        yield self._record(level='debug', msg='Starting gti check')
-
-        try:
-            gti = self.obs.load('gti')
-        except:
-            yield self._record(level='warning', msg='Loading GTI failed')
-            return
-
-        if len(gti.table) == 0:
-            yield self._record(level='error', msg='GTI table has zero rows')
-
-        columns_required = ['START', 'STOP']
-        columns_missing = set(columns_required) - set(gti.table.colnames)
-        if columns_missing:
-            yield self._record(level='error', msg='Missing table columns: {!r}'.format(columns_missing))
-
-        # TODO: Check that header keywords agree with table entries
-        # TSTART, TSTOP, MJDREFI, MJDREFF
-
-        # Check that START and STOP times are consecutive
-        # times = np.ravel(self.table['START'], self.table['STOP'])
-        # # TODO: not sure this is correct ... add test with a multi-gti table from Fermi.
-        # if not np.all(np.diff(times) >= 0):
-        #     yield 'GTIs are not consecutive or sorted.'
-
-    def _check_times(self):
-        """Check if various times are consistent.
-
-        The headers and tables of the FITS EVENTS and GTI extension
-        contain various observation and event time information.
-        """
-        ok = True
-
-        # http://fermi.gsfc.nasa.gov/ssc/data/analysis/documentation/Cicerone/Cicerone_Data/Time_in_ScienceTools.html
-        # https://hess-confluence.desy.de/confluence/display/HESS/HESS+FITS+data+-+References+and+checks#HESSFITSdata-Referencesandchecks-Time
-        telescope_met_refs = OrderedDict(
-            FERMI=Time('2001-01-01T00:00:00'),
-            HESS=Time('2001-01-01T00:00:00'),
-        )
-
-        meta = self.dset.event_list.table.meta
-        telescope = meta['TELESCOP']
-
-        if telescope in telescope_met_refs.keys():
-            dt = (self.time_ref - telescope_met_refs[telescope])
-            if dt > self.accuracy['time']:
-                ok = False
-                self.logger.error('MET reference is incorrect.')
-        else:
-            self.logger.debug('Skipping MET reference check ... not known for this telescope.')
-
-        # TODO: emit warning if event times aren't time-ordered.
-
-        # TODO: check latest CTA spec to see which info is required / optional
-        # EVENTS header keywords:
-        # 'DATE_OBS': '2004-10-14'
-        # 'TIME_OBS': '00:08:27'
-        # 'DATE_END': '2004-10-14'
-        # 'TIME_END': '00:34:44'
-        # 'TSTART': 150984507.0
-        # 'TSTOP': 150986084.0
-        # 'MJDREFI': 51544
-        # 'MJDREFF': 0.5
-        # 'TIMEUNIT': 's'
-        # 'TIMESYS': 'TT'
-        # 'TIMEREF': 'local'
-        # 'TASSIGN': 'Namibia'
-        # 'TELAPSE': 0
-        # 'ONTIME': 1577.0
-        # 'LIVETIME': 1510.95910644531
-        # 'DEADC': 0.964236799627542
-
-        return ok
-
-
-    def check_aeff(self):
-        yield self._record(level='debug', msg='Starting aeff check')
-
-        try:
-            aeff = self.obs.load('aeff')
-        except:
-            yield self._record(level='warning', msg='Loading aeff failed')
-            return
-
-        # Check that thresholds are meaningful for aeff
-        if 'LO_THRES' in aeff.meta and 'HI_THRES' in aeff.meta and aeff.meta['LO_THRES'] >= aeff.meta['HI_THRES']:
-            yield self._record(level='error', msg='LO_THRES >= HI_THRES in effective area meta data')
-
-        # Check that maximum value of aeff is greater than zero
-        if np.max(aeff.data.data) <= 0:
-            yield self._record(level='error', msg='maximum entry of effective area table <= 0')
-
-    def check_edisp(self):
-        yield self._record(level='debug', msg='Starting edisp check')
-
-        try:
-            edisp = self.obs.load('edisp')
-        except:
-            yield self._record(level='warning', msg='Loading edisp failed')
-            return
-
-        # Check that maximum value of edisp matrix is greater than zero
-        if np.max(edisp.data.data) <= 0:
-            yield self._record(level='error', msg='maximum entry of edisp is <= 0')
-
-    def check_psf(self):
-        yield self._record(level='debug', msg='Starting psf check')
-
-        try:
-            psf = self.obs.load('psf')
-        except:
-            yield self._record(level='warning', msg='Loading psf failed')
-            return
